Fix compilation on Fedora 23 i686

+++ api/server/credentials_linux.go
@@ -213,7 +213,7 @@ func (s *Server) LogAction(w http.ResponseWriter, r *http.Request) error {
 			break
 		}
 		message = fmt.Sprintf("LoginUID=%v, %s", loginuid, message)
-		if loginuid == 0xffffffff { // -1 means no login user
+		if uint32(loginuid) == 0xffffffff { // -1 means no login user
 			//No login UID is set, so no point in looking up a name
 			break
 		}
